+++ widgets/text/options.go
@@ -17,6 +17,7 @@ package text
 import (
 	"fmt"
 
+	"github.com/mum4k/termdash/internal/wrap"
 	"github.com/mum4k/termdash/keyboard"
 	"github.com/mum4k/termdash/mouse"
 )
@@ -31,7 +32,7 @@ type Option interface {
 
 // options stores the provided options.
 type options struct {
-	wrapAtRunes      bool
+	wrapMode         wrap.Mode
 	rollContent      bool
 	disableScrolling bool
 	mouseUpButton    mouse.Button
@@ -83,12 +84,21 @@ func (o option) set(opts *options) {
 	o(opts)
 }
 
+// WrapAtWords configures the text widget so that it automatically wraps lines
+// that are longer than the width of the widget at word boundaries. If not
+// provided, long lines are trimmed instead.
+func WrapAtWords() Option {
+	return option(func(opts *options) {
+		opts.wrapMode = wrap.AtWords
+	})
+}
+
 // WrapAtRunes configures the text widget so that it automatically wraps lines
 // that are longer than the width of the widget at rune boundaries. If not
 // provided, long lines are trimmed instead.
 func WrapAtRunes() Option {
 	return option(func(opts *options) {
-		opts.wrapAtRunes = true
+		opts.wrapMode = wrap.AtRunes
 	})
 }